+++ example/example/spiders/books_04_overrides_01.py
@@ -6,7 +6,7 @@ the crawling logic (the spider is exactly the same).
 The default configured PO logic contains the logic for books.toscrape.com
 """
 import scrapy
-from web_poet import WebPage
+from web_poet import ApplyRule, WebPage
 
 from scrapy_poet import callback_for
 
@@ -51,8 +51,8 @@ class BooksSpider(scrapy.Spider):
     # Configuring different page objects pages from the bookpage.com domain
     custom_settings = {
         "SCRAPY_POET_OVERRIDES": [
-            ("bookpage.com", BPBookListPage, BookListPage),
-            ("bookpage.com", BPBookPage, BookPage),
+            ApplyRule("bookpage.com", use=BPBookListPage, instead_of=BookListPage),
+            ApplyRule("bookpage.com", use=BPBookPage, instead_of=BookPage),
         ]
     }